Work status:10/03/2019

+++ WinCDS/Forms/ArCard.vb
@@ -750,12 +750,12 @@ ErrorHandler:
         DisposeDA(R)
     End Function
 
-    Public Function QueryPayLogSale(ByVal I As Long) As String
+    Public Function QueryPayLogSale(ByVal I As Integer) As String
         If I > PayCount Then Exit Function
         QueryPayLogSale = PayLog(I - 1).SaleNo
     End Function
 
-    Public Function QueryPayLogAmount(ByVal I As Long) As Decimal
+    Public Function QueryPayLogAmount(ByVal I As Integer) As Decimal
         If I > PayCount Then Exit Function
         QueryPayLogAmount = PayLog(I - 1).Amount
     End Function
